Backend/requirements.txt
# ── Existing scanner dependencies ────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.5.1
pymongo==4.8.0
pydantic==2.9.2
pydantic[email]==2.9.2
email-validator==2.2.0
pydantic-settings==2.5.2
python-dotenv==1.0.1
cryptography==43.0.1
python-nmap==0.7.1
httpx==0.27.2

# ── PostgreSQL + SQLAlchemy (async) ──────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.13.3

# ── Auth ─────────────────────────────────────────────────────────
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.12

# ── Rate limiting ────────────────────────────────────────────────
slowapi==0.1.9

# ── Report generation ────────────────────────────────────────────
reportlab==4.2.5
